fix:Replaced Opportunity creation with mapped doc
